Move on modifier docs to @ember/helpers keyword section
